Agents/requirements.txt
requests == 2.24.0
paho-mqtt == 1.5.1
jsonify == 0.5
websocket_client == 0.57.0
psutil == 5.7.2
